requirements.txt
Django==1.5.2
# MySQL-python==1.2.3
mysqlclient
Pillow==2.5.3
PyBrain==0.3
South==1.0
TornadIO2==0.0.4
argparse==1.2.1
async==0.6.1
certifi
chardet
django-classy-tags==0.5.1
django-less==0.7.2
django-mptt==0.5.3
django-robokassa==1.2
django-sekizai==0.7
djangorestframework==2.4.4
gitdb==0.5.4
gunicorn==19.1.1
lesscpy==0.10.1
lesscss==0.1.4
medusa==0.5.4
meld3==0.6.10
ping==0.2
ply==3.4
pymorphy==0.5.6
redis==2.10.3
requests
simplejson==3.6.3
smmap==0.8.2
sorl-thumbnail==11.12.1b
supervisor==3.0a8
tornado==4.3
scipy
tornado-redis==2.4.1
virtualenv==1.11.6
wsgiref==0.1.2
model_mommy==1.3.2
robot
robotframework==2.8.7
robotframework-selenium2library==1.6.0
maildump
django-pipeline==1.3.24
docx2html==0.2.3
xlsxwriter
openpyxl==2.3.1
lxml==3.5
django-yandex-money
django-notify
django-environ
mock

